Codegen and update types and returned shapes
(cherry picked from commit 6e6b00627e65bb05f3dca6a63bc777bea89b2393)
