+++ lexicons/app/bsky/unspecced/getPostThreadHiddenV2.json
@@ -29,46 +29,13 @@
           "properties": {
             "thread": {
               "type": "array",
-              "description": "A flat list of thread hidden items. The depth of each item is indicated by the depth property inside the item.",
+              "description": "A flat list of hidden thread items. The depth of each item is indicated by the depth property inside the item.",
               "items": {
                 "type": "ref",
-                "ref": "#threadHiddenItem"
+                "ref": "app.bsky.unspecced.defs#threadItem"
               }
             }
           }
-        }
-      }
-    },
-    "threadHiddenItem": {
-      "type": "object",
-      "required": ["uri", "depth", "value"],
-      "properties": {
-        "uri": {
-          "type": "string",
-          "format": "at-uri"
-        },
-        "depth": {
-          "type": "integer",
-          "description": "The nesting level of this item in the thread. Depth 0 means the anchor item. Items above have negative depths, items below have positive depths."
-        },
-        "value": {
-          "type": "union",
-          "refs": ["#threadHiddenItemPost"]
-        }
-      }
-    },
-    "threadHiddenItemPost": {
-      "type": "object",
-      "required": ["post", "hiddenByThreadgate", "mutedByViewer"],
-      "properties": {
-        "post": { "type": "ref", "ref": "app.bsky.feed.defs#postView" },
-        "hiddenByThreadgate": {
-          "type": "boolean",
-          "description": "The threadgate created by the author indicates this post as a reply to be hidden for everyone consuming the thread."
-        },
-        "mutedByViewer": {
-          "type": "boolean",
-          "description": "This is by an account muted by the viewer requesting it."
         }
       }
     }
