+++ lexicons/app/bsky/unspecced/getPostThreadV2.json
@@ -57,7 +57,7 @@
               "description": "A flat list of thread items. The depth of each item is indicated by the depth property inside the item.",
               "items": {
                 "type": "ref",
-                "ref": "#threadItem"
+                "ref": "app.bsky.unspecced.defs#threadItem"
               }
             },
             "threadgate": {
@@ -70,63 +70,6 @@
             }
           }
         }
-      }
-    },
-    "threadItem": {
-      "type": "object",
-      "required": ["uri", "depth", "value"],
-      "properties": {
-        "uri": {
-          "type": "string",
-          "format": "at-uri"
-        },
-        "depth": {
-          "type": "integer",
-          "description": "The nesting level of this item in the thread. Depth 0 means the anchor item. Items above have negative depths, items below have positive depths."
-        },
-        "value": {
-          "type": "union",
-          "refs": [
-            "#threadItemPost",
-            "#threadItemNoUnauthenticated",
-            "#threadItemNotFound",
-            "#threadItemBlocked"
-          ]
-        }
-      }
-    },
-    "threadItemPost": {
-      "type": "object",
-      "required": ["post", "moreParents", "moreReplies", "opThread"],
-      "properties": {
-        "post": { "type": "ref", "ref": "app.bsky.feed.defs#postView" },
-        "moreParents": {
-          "type": "boolean",
-          "description": "This post has more parents that were not present in the response. This is just a boolean, without the number of parents."
-        },
-        "moreReplies": {
-          "type": "integer",
-          "description": "This post has more replies that were not present in the response. This is a numeric value, which is best-effort and might not be accurate."
-        },
-        "opThread": {
-          "type": "boolean",
-          "description": "This post is part of a contiguous thread by the OP from the thread root. Many different OP threads can happen in the same thread."
-        }
-      }
-    },
-    "threadItemNoUnauthenticated": {
-      "type": "object",
-      "properties": {}
-    },
-    "threadItemNotFound": {
-      "type": "object",
-      "properties": {}
-    },
-    "threadItemBlocked": {
-      "type": "object",
-      "required": ["author"],
-      "properties": {
-        "author": { "type": "ref", "ref": "app.bsky.feed.defs#blockedAuthor" }
       }
     }
   }
