+++ packages/api/src/client/types/app/bsky/unspecced/getPostThreadHiddenV2.ts
@@ -10,7 +10,7 @@ import {
   is$typed as _is$typed,
   type OmitKey,
 } from '../../../../util'
-import type * as AppBskyFeedDefs from '../feed/defs.js'
+import type * as AppBskyUnspeccedDefs from './defs.js'
 
 const is$typed = _is$typed,
   validate = _validate
@@ -26,8 +26,8 @@ export interface QueryParams {
 export type InputSchema = undefined
 
 export interface OutputSchema {
-  /** A flat list of thread hidden items. The depth of each item is indicated by the depth property inside the item. */
-  thread: ThreadHiddenItem[]
+  /** A flat list of hidden thread items. The depth of each item is indicated by the depth property inside the item. */
+  thread: AppBskyUnspeccedDefs.ThreadItem[]
 }
 
 export interface CallOptions {
@@ -43,41 +43,4 @@ export interface Response {
 
 export function toKnownErr(e: any) {
   return e
-}
-
-export interface ThreadHiddenItem {
-  $type?: 'app.bsky.unspecced.getPostThreadHiddenV2#threadHiddenItem'
-  uri: string
-  /** The nesting level of this item in the thread. Depth 0 means the anchor item. Items above have negative depths, items below have positive depths. */
-  depth: number
-  value: $Typed<ThreadHiddenItemPost> | { $type: string }
-}
-
-const hashThreadHiddenItem = 'threadHiddenItem'
-
-export function isThreadHiddenItem<V>(v: V) {
-  return is$typed(v, id, hashThreadHiddenItem)
-}
-
-export function validateThreadHiddenItem<V>(v: V) {
-  return validate<ThreadHiddenItem & V>(v, id, hashThreadHiddenItem)
-}
-
-export interface ThreadHiddenItemPost {
-  $type?: 'app.bsky.unspecced.getPostThreadHiddenV2#threadHiddenItemPost'
-  post: AppBskyFeedDefs.PostView
-  /** The threadgate created by the author indicates this post as a reply to be hidden for everyone consuming the thread. */
-  hiddenByThreadgate: boolean
-  /** This is by an account muted by the viewer requesting it. */
-  mutedByViewer: boolean
-}
-
-const hashThreadHiddenItemPost = 'threadHiddenItemPost'
-
-export function isThreadHiddenItemPost<V>(v: V) {
-  return is$typed(v, id, hashThreadHiddenItemPost)
-}
-
-export function validateThreadHiddenItemPost<V>(v: V) {
-  return validate<ThreadHiddenItemPost & V>(v, id, hashThreadHiddenItemPost)
 }
