+++ packages/bsky/src/lexicon/types/app/bsky/unspecced/getPostThreadV2.ts
@@ -11,6 +11,7 @@ import {
   type OmitKey,
 } from '../../../../util'
 import { HandlerAuth, HandlerPipeThrough } from '@atproto/xrpc-server'
+import type * as AppBskyUnspeccedDefs from './defs.js'
 import type * as AppBskyFeedDefs from '../feed/defs.js'
 
 const is$typed = _is$typed,
@@ -36,7 +37,7 @@ export type InputSchema = undefined
 
 export interface OutputSchema {
   /** A flat list of thread items. The depth of each item is indicated by the depth property inside the item. */
-  thread: ThreadItem[]
+  thread: AppBskyUnspeccedDefs.ThreadItem[]
   threadgate?: AppBskyFeedDefs.ThreadgateView
   /** Whether this thread has hidden replies. If true, a call can be made to the `getPostThreadHiddenV2` endpoint to retrieve them. */
   hasHiddenReplies: boolean
@@ -67,94 +68,3 @@ export type HandlerReqCtx<HA extends HandlerAuth = never> = {
 export type Handler<HA extends HandlerAuth = never> = (
   ctx: HandlerReqCtx<HA>,
 ) => Promise<HandlerOutput> | HandlerOutput
-
-export interface ThreadItem {
-  $type?: 'app.bsky.unspecced.getPostThreadV2#threadItem'
-  uri: string
-  /** The nesting level of this item in the thread. Depth 0 means the anchor item. Items above have negative depths, items below have positive depths. */
-  depth: number
-  value:
-    | $Typed<ThreadItemPost>
-    | $Typed<ThreadItemNoUnauthenticated>
-    | $Typed<ThreadItemNotFound>
-    | $Typed<ThreadItemBlocked>
-    | { $type: string }
-}
-
-const hashThreadItem = 'threadItem'
-
-export function isThreadItem<V>(v: V) {
-  return is$typed(v, id, hashThreadItem)
-}
-
-export function validateThreadItem<V>(v: V) {
-  return validate<ThreadItem & V>(v, id, hashThreadItem)
-}
-
-export interface ThreadItemPost {
-  $type?: 'app.bsky.unspecced.getPostThreadV2#threadItemPost'
-  post: AppBskyFeedDefs.PostView
-  /** This post has more parents that were not present in the response. This is just a boolean, without the number of parents. */
-  moreParents: boolean
-  /** This post has more replies that were not present in the response. This is a numeric value, which is best-effort and might not be accurate. */
-  moreReplies: number
-  /** This post is part of a contiguous thread by the OP from the thread root. Many different OP threads can happen in the same thread. */
-  opThread: boolean
-}
-
-const hashThreadItemPost = 'threadItemPost'
-
-export function isThreadItemPost<V>(v: V) {
-  return is$typed(v, id, hashThreadItemPost)
-}
-
-export function validateThreadItemPost<V>(v: V) {
-  return validate<ThreadItemPost & V>(v, id, hashThreadItemPost)
-}
-
-export interface ThreadItemNoUnauthenticated {
-  $type?: 'app.bsky.unspecced.getPostThreadV2#threadItemNoUnauthenticated'
-}
-
-const hashThreadItemNoUnauthenticated = 'threadItemNoUnauthenticated'
-
-export function isThreadItemNoUnauthenticated<V>(v: V) {
-  return is$typed(v, id, hashThreadItemNoUnauthenticated)
-}
-
-export function validateThreadItemNoUnauthenticated<V>(v: V) {
-  return validate<ThreadItemNoUnauthenticated & V>(
-    v,
-    id,
-    hashThreadItemNoUnauthenticated,
-  )
-}
-
-export interface ThreadItemNotFound {
-  $type?: 'app.bsky.unspecced.getPostThreadV2#threadItemNotFound'
-}
-
-const hashThreadItemNotFound = 'threadItemNotFound'
-
-export function isThreadItemNotFound<V>(v: V) {
-  return is$typed(v, id, hashThreadItemNotFound)
-}
-
-export function validateThreadItemNotFound<V>(v: V) {
-  return validate<ThreadItemNotFound & V>(v, id, hashThreadItemNotFound)
-}
-
-export interface ThreadItemBlocked {
-  $type?: 'app.bsky.unspecced.getPostThreadV2#threadItemBlocked'
-  author: AppBskyFeedDefs.BlockedAuthor
-}
-
-const hashThreadItemBlocked = 'threadItemBlocked'
-
-export function isThreadItemBlocked<V>(v: V) {
-  return is$typed(v, id, hashThreadItemBlocked)
-}
-
-export function validateThreadItemBlocked<V>(v: V) {
-  return validate<ThreadItemBlocked & V>(v, id, hashThreadItemBlocked)
-}
