+++ packages/pds/src/lexicon/types/app/bsky/unspecced/defs.ts
@@ -10,6 +10,7 @@ import {
   type OmitKey,
 } from '../../../../util'
 import type * as AppBskyActorDefs from '../actor/defs.js'
+import type * as AppBskyFeedDefs from '../feed/defs.js'
 
 const is$typed = _is$typed,
   validate = _validate
@@ -124,4 +125,99 @@ export function isTrendView<V>(v: V) {
 
 export function validateTrendView<V>(v: V) {
   return validate<TrendView & V>(v, id, hashTrendView)
+}
+
+export interface ThreadItem {
+  $type?: 'app.bsky.unspecced.defs#threadItem'
+  uri: string
+  /** The nesting level of this item in the thread. Depth 0 means the anchor item. Items above have negative depths, items below have positive depths. */
+  depth: number
+  value:
+    | $Typed<ThreadItemPost>
+    | $Typed<ThreadItemNoUnauthenticated>
+    | $Typed<ThreadItemNotFound>
+    | $Typed<ThreadItemBlocked>
+    | { $type: string }
+}
+
+const hashThreadItem = 'threadItem'
+
+export function isThreadItem<V>(v: V) {
+  return is$typed(v, id, hashThreadItem)
+}
+
+export function validateThreadItem<V>(v: V) {
+  return validate<ThreadItem & V>(v, id, hashThreadItem)
+}
+
+export interface ThreadItemPost {
+  $type?: 'app.bsky.unspecced.defs#threadItemPost'
+  post: AppBskyFeedDefs.PostView
+  /** This post has more parents that were not present in the response. This is just a boolean, without the number of parents. */
+  moreParents: boolean
+  /** This post has more replies that were not present in the response. This is a numeric value, which is best-effort and might not be accurate. */
+  moreReplies: number
+  /** This post is part of a contiguous thread by the OP from the thread root. Many different OP threads can happen in the same thread. */
+  opThread: boolean
+  /** The threadgate created by the author indicates this post as a reply to be hidden for everyone consuming the thread. */
+  hiddenByThreadgate: boolean
+  /** This is by an account muted by the viewer requesting it. */
+  mutedByViewer: boolean
+}
+
+const hashThreadItemPost = 'threadItemPost'
+
+export function isThreadItemPost<V>(v: V) {
+  return is$typed(v, id, hashThreadItemPost)
+}
+
+export function validateThreadItemPost<V>(v: V) {
+  return validate<ThreadItemPost & V>(v, id, hashThreadItemPost)
+}
+
+export interface ThreadItemNoUnauthenticated {
+  $type?: 'app.bsky.unspecced.defs#threadItemNoUnauthenticated'
+}
+
+const hashThreadItemNoUnauthenticated = 'threadItemNoUnauthenticated'
+
+export function isThreadItemNoUnauthenticated<V>(v: V) {
+  return is$typed(v, id, hashThreadItemNoUnauthenticated)
+}
+
+export function validateThreadItemNoUnauthenticated<V>(v: V) {
+  return validate<ThreadItemNoUnauthenticated & V>(
+    v,
+    id,
+    hashThreadItemNoUnauthenticated,
+  )
+}
+
+export interface ThreadItemNotFound {
+  $type?: 'app.bsky.unspecced.defs#threadItemNotFound'
+}
+
+const hashThreadItemNotFound = 'threadItemNotFound'
+
+export function isThreadItemNotFound<V>(v: V) {
+  return is$typed(v, id, hashThreadItemNotFound)
+}
+
+export function validateThreadItemNotFound<V>(v: V) {
+  return validate<ThreadItemNotFound & V>(v, id, hashThreadItemNotFound)
+}
+
+export interface ThreadItemBlocked {
+  $type?: 'app.bsky.unspecced.defs#threadItemBlocked'
+  author: AppBskyFeedDefs.BlockedAuthor
+}
+
+const hashThreadItemBlocked = 'threadItemBlocked'
+
+export function isThreadItemBlocked<V>(v: V) {
+  return is$typed(v, id, hashThreadItemBlocked)
+}
+
+export function validateThreadItemBlocked<V>(v: V) {
+  return validate<ThreadItemBlocked & V>(v, id, hashThreadItemBlocked)
 }
